refactor: send_notify_to_host.ts を common/ に移動
