+++ src/app/feat-recruit/cancel_request/cancel_notify_event.ts
@@ -6,8 +6,8 @@ import { recoveryThinkingButton } from '../../common/button_components';
 import { assertExistCheck } from '../../common/others';
 import { sendStickyMessage } from '../../common/sticky_message';
 import { StickyKey } from '../../constant/sticky_key';
-import { memberListText } from '../interactions/buttons/other_events';
-import { sendCancelNotifyToHost } from '../interactions/buttons/send_notify_to_host';
+import { memberListText } from '../common/member_list';
+import { sendCancelNotifyToHost } from '../common/send_notify_to_host';
 import { availableRecruitString } from '../sticky/recruit_sticky_messages';
 
 export async function cancelRequestNotify(
